Original commit
This is the starting point at this point the camera is still not working and is reading and writing to a file because we do not have a database yet. There is also another project that I have created to make testing the reading and writing to a file easier.